backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
aiofiles==24.1.0
email-validator==2.2.0
pdfminer.six==20221105
httpx==0.27.0
